int nHeaderSize -> size_t nHeaderSize.
gdalpamproxydb.cpp:108:57: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
gdalpamproxydb.cpp:224:58: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]


git-svn-id: https://svn.osgeo.org/gdal/trunk@33766 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/gcore/gdalpamproxydb.cpp
@@ -102,7 +102,7 @@ void GDALPamProxyDB::LoadDB()
 /* -------------------------------------------------------------------- */
 /*      Read header, verify and extract update counter.                 */
 /* -------------------------------------------------------------------- */
-    const int nHeaderSize = 100;
+    const size_t nHeaderSize = 100;
     GByte abyHeader[nHeaderSize] = { '\0' };
 
     if( VSIFReadL( abyHeader, 1, nHeaderSize, fpDB ) != nHeaderSize
@@ -213,7 +213,7 @@ void GDALPamProxyDB::SaveDB()
 /* -------------------------------------------------------------------- */
 /*      Write header.                                                   */
 /* -------------------------------------------------------------------- */
-    const int nHeaderSize = 100;
+    const size_t nHeaderSize = 100;
     GByte abyHeader[nHeaderSize] = { '\0' };
 
     memset( abyHeader, ' ', sizeof(abyHeader) );
